Report for each Property should working
- Delete Settings from the template.
- Udpdated all properties where was used float type and need more precision to decimal.
- Created Background services to run by schedule.

+++ src/Web/EnergySystem.Web.ViewModels/Battery/BaseBatteryInputModel.cs
@@ -47,7 +47,7 @@
         public float Temperature { get; set; }
 
         [Required]
-        public float LifetimeEnergyStored { get; set; }
+        public decimal LifetimeEnergyStored { get; set; }
 
         [Required]
         public string PropertyId { get; set; }
